fixing storybook circular dependency
