+++ helpers/multiply/calculations.test.ts
@@ -1,8 +1,4 @@
 import BigNumber from 'bignumber.js'
-import { expect } from 'chai'
-import { mockManageMultiplyVault$ } from 'helpers/mocks/manageMultiplyVault.mock'
-import { calculatePNL } from 'helpers/multiply/calculations'
-import { getStateUnpacker } from 'helpers/testHelpers'
 import { one, zero } from 'helpers/zero'
 
 import { VaultHistoryEvent } from '../../features/vaultHistory/vaultHistory'
@@ -125,26 +121,3 @@ export const mockedMultiplyEvents: VaultHistoryEvent[] = [
     gasFee: new BigNumber(0.02825),
   },
 ]
-
-describe('Multiply calculations', () => {
-  it('Calculates PNL correctly', () => {
-    const state = getStateUnpacker(
-      mockManageMultiplyVault$({
-        vault: {
-          collateral: new BigNumber('20'),
-          debt: new BigNumber('16680'),
-        },
-        priceInfo: {
-          ethPrice: new BigNumber('2650'),
-          collateralPrice: new BigNumber('2650'),
-        },
-        exchangeQuote: {
-          marketPrice: new BigNumber('2650'),
-        },
-      }),
-    )
-
-    const pnl = calculatePNL(mockedMultiplyEvents, state().netValueUSD)
-    expect(pnl.decimalPlaces(20)).to.be.deep.equal(new BigNumber('0.27552238805970149253'))
-  })
-})
